refacted filter step
